Extract Converter creation to ConverterPool class
