+++ lib/lebab.js
@@ -1,13 +1,14 @@
 "use babel";
-import Converter from './Converter';
+import ConverterPool from './ConverterPool';
 import transforms from './transforms';
 import config from './config';
 
 export default {
     config,
-    converterMap: new WeakMap(),
 
     activate() {
+        this.converterPool = new ConverterPool();
+
         // Command for applying all enabled transforms
         atom.commands.add(
             "atom-workspace",
@@ -26,36 +27,6 @@ export default {
     },
 
     convert(transforms) {
-        this.getConverter().convert(transforms);
-    },
-
-    getConverter() {
-        const editor = atom.workspace.getActiveTextEditor();
-        if (!this.converterMap.get(editor)) {
-            this.converterMap.set(editor, this.createConverter(editor));
-        }
-        return this.converterMap.get(editor);
-    },
-
-    createConverter(editor) {
-        const converter = new Converter(editor);
-
-        // Show/Hide warnings panel when active editor changes
-        const listener = atom.workspace.onDidStopChangingActivePaneItem(() => {
-            if (atom.workspace.getActiveTextEditor() === editor) {
-               converter.getWarningsPanel().show();
-            }
-            else {
-               converter.getWarningsPanel().hide();
-            }
-        });
-
-        // When editor is closed, eliminate warnings panel
-        editor.onDidDestroy(() => {
-            converter.getWarningsPanel().destroy();
-            listener.dispose();
-        });
-
-        return converter;
+        this.converterPool.getActive().convert(transforms);
     },
 }
